refactor: bootstrap transformer moved to transformers
