+++ index.js
@@ -10,6 +10,7 @@ const {
 
 Object.assign(exports, require('./plugins'));
 Object.assign(exports, require('./host/resolver'));
+Object.assign(exports, require('./transformers/ns-replace-bootstrap'));
 
 exports.getAotEntryModule = function (appDirectory) {
 	verifyEntryModuleDirectory(appDirectory);
